Add relevant modules to the project
